Actually make a CDF
The cumulative=-1 option actually generates a quantile function, or an
inverse cumulative distribution function. The regular cumulative
distribution function is most clearly described as [1]

    the probability that X will take a value less than or equal to x

whereas the quantile function [2]

    tells you what x would make F(x) return some value p

In summary, the CDF translates x to p and the quantile function
translates p to x.

Because this generally moves the knee to the top left, force the legend
to the lower right.

[1] https://en.wikipedia.org/wiki/Cumulative_distribution_function
[2] https://stats.stackexchange.com/a/212828